refactor - break out opFns
